[Bronze II] Title: 숫자, Time: 12 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -7,7 +7,7 @@ int main()
     ios::sync_with_stdio(false);
     cin.tie(nullptr);
     
-    int A, B, cnt;
+    long long A, B, cnt;
     cin >> A >> B;
     
     if (A == B)
